# api/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "TAMS API",
    "description": "Time-addressable Media Store API - A comprehensive FastAPI implementation of the BBC TAMS API specification with VAST database integration for high-performance time-series analytics and S3-compatible storage for media segments.",
    "version": "6.0",
    "contact": {
      "name": "TAMS API Support",
      "url": "https://github.com/jesseVast/vasttams"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.tams.example.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/analytics/flow-usage": {
      "get": {
        "summary": "Get flow usage analytics",
        "operationId": "get_flow_usage_analytics",
        "tags": [
          "Analytics"
        ],
        "responses": {
          "200": {
            "description": "Flow usage analytics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total_flows": {
                      "type": "integer"
                    },
                    "format_distribution": {
                      "type": "object"
                    },
                    "estimated_storage_bytes": {
                      "type": "integer"
                    },
                    "average_flow_size": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/analytics/storage-usage": {
      "get": {
        "summary": "Get storage usage analytics",
        "operationId": "get_storage_usage_analytics",
        "tags": [
          "Analytics"
        ],
        "responses": {
          "200": {
            "description": "Storage usage analytics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total_objects": {
                      "type": "integer"
                    },
                    "total_size_bytes": {
                      "type": "integer"
                    },
                    "average_size_bytes": {
                      "type": "number"
                    },
                    "most_accessed": {
                      "type": "integer"
                    },
                    "least_accessed": {
                      "type": "integer"
                    },
                    "average_access_count": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/analytics/time-range-analysis": {
      "get": {
        "summary": "Get time range analytics",
        "operationId": "get_time_range_analytics",
        "tags": [
          "Analytics"
        ],
        "responses": {
          "200": {
            "description": "Time range analytics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total_segments": {
                      "type": "integer"
                    },
                    "average_duration_seconds": {
                      "type": "number"
                    },
                    "min_duration_seconds": {
                      "type": "number"
                    },
                    "max_duration_seconds": {
                      "type": "number"
                    },
                    "total_duration_seconds": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}